refactor: Implement custom hooks for context menu and keyboard navigation; enhance team grouping logic
